Move the call to Deploy out of createOrUpate
Also try and simplify the initializer/updater interface

This is only part 1 of #390
