+++ pkg/upgrade/types.go
@@ -6,17 +6,17 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/openshift/openshift-azure/pkg/api"
-	"github.com/openshift/openshift-azure/pkg/initialize"
 	"github.com/openshift/openshift-azure/pkg/log"
 )
 
 type Upgrader interface {
-	initialize.Initializer
-	Update(ctx context.Context, cs *api.OpenShiftManagedCluster, azuredeploy []byte, config api.PluginConfig) error
+	InitializeCluster(ctx context.Context, cs *api.OpenShiftManagedCluster) error
+	Deploy(ctx context.Context, cs *api.OpenShiftManagedCluster, azuredeploy []byte) error
+	Update(ctx context.Context, cs *api.OpenShiftManagedCluster, azuredeploy []byte) error
 }
 
 type simpleUpgrader struct {
-	initialize.Initializer
+	pluginConfig api.PluginConfig
 }
 
 var _ Upgrader = &simpleUpgrader{}
@@ -24,6 +24,6 @@ var _ Upgrader = &simpleUpgrader{}
 func NewSimpleUpgrader(entry *logrus.Entry, pluginConfig api.PluginConfig) Upgrader {
 	log.New(entry)
 	return &simpleUpgrader{
-		Initializer: initialize.NewSimpleInitializer(entry, pluginConfig),
+		pluginConfig: pluginConfig,
 	}
 }
